requirements.txt
marqo==3.1.0
streamlit
nltk==3.8.1
numpy==1.23.5
pandas==1.5.3
scikit_learn==1.3.0
umap_learn==0.5.4
datashader
matplotlib
colorcet
bokeh
holoviews
scikit-image
plotly
seaborn